refactor: ErrorCode enum 이름을 ResponseCode로 변경
