+++ src/main/java/com/spartaschedulerdevelop/service/LoginService.java
@@ -1,8 +1,8 @@
 package com.spartaschedulerdevelop.service;
 
 import com.spartaschedulerdevelop.common.config.PasswordEncoder;
-import com.spartaschedulerdevelop.common.exception.ErrorCode;
-import com.spartaschedulerdevelop.common.exception.MyCustomException;
+import com.spartaschedulerdevelop.common.advice.ResponseCode;
+import com.spartaschedulerdevelop.common.advice.exception.MyCustomException;
 import com.spartaschedulerdevelop.dto.login.LoginRequestDto;
 import com.spartaschedulerdevelop.dto.login.LoginResponseDto;
 import com.spartaschedulerdevelop.entity.User;
@@ -20,10 +20,10 @@ public class LoginService {
     private final PasswordEncoder passwordEncoder;
 
     public LoginResponseDto authenticate(LoginRequestDto request) {
-        User user = userRepository.findByEmail(request.email()).orElseThrow(() -> new MyCustomException(ErrorCode.INVALID_CREDENTIALS));
+        User user = userRepository.findByEmail(request.email()).orElseThrow(() -> new MyCustomException(ResponseCode.INVALID_CREDENTIALS));
 
         if(!passwordEncoder.matches(request.password(), user.getPassword())){
-            throw new MyCustomException(ErrorCode.INVALID_CREDENTIALS);
+            throw new MyCustomException(ResponseCode.INVALID_CREDENTIALS);
         }
 
         return loginMapper.toLoginResponseDto(user);
